Fixed typo

+++ src/core/Aggregators.java
@@ -589,7 +589,7 @@ public final class Aggregators {
       long total = 0;
       long numOnes = 0;
       do {
-        final double x = values.nextLongValue();
+        final double x = values.nextDoubleValue();
         total++;
         if (x == 1) numOnes++;
       } while (values.hasNextValue());
@@ -630,7 +630,7 @@ public final class Aggregators {
       long total = 0;
       long numTwos = 0;
       do {
-        final double x = values.nextLongValue();
+        final double x = values.nextDoubleValue();
         total++;
         if (x == 1) numTwos++;
       } while (values.hasNextValue());
